TypeScript migration 1
