+++ src/utils/dom.ts
@@ -1,17 +1,13 @@
 /**
  * 주어진 element의 자식들을 모두 탐색합니다.
- * 
- * @param {HTMLElement} element 탐색할 element.
+ *
+ * @param element 탐색할 element.
  */
-const traversal = element => {
+export const traversal = (element: HTMLElement): HTMLElement[] => {
   let result = []
 
-  if (!element instanceof HTMLElement) {
-    throw new Error('Given argument is not a HTMLElement.')
-  }
-
   if (element.nodeType !== Node.ELEMENT_NODE) {
-    return false
+    return []
   }
 
   let childs = element.children
@@ -20,17 +16,13 @@ const traversal = element => {
   result.push(element)
 
   for (var i = 0; i < child_len; i++) {
-    let child = childs[i]
+    let child = childs[i] as HTMLElement
 
     let travel = traversal(child)
-    if (travel) {
+    if (travel.length) {
       result.push(...travel)
     }
   }
 
   return result
-}
-
-module.exports = {
-  traversal
 }
